Parametrize tests, add generic QA test (#580)
This commit adds testing for Bento operation on staged file
modifications.

To make this a bit easier to write, I've updated qa.py in a couple of
areas:
- I use parameterized pytest to avoid having to update qa.py when we
  add or rename QA tests
- I also remove git hashes when we are operating on a git master branch
- You can now enter commands all on one line
- You don't need to check out a repo to run a QA test
- You can reset text for one test only by running:
    pythong tests/acceptance/qa.py TEST_NAME

I also remove duplicated generic testing from the repo-specific tests,
as this was just a waste of time.

+++ tests/acceptance/qa.py
@@ -1,6 +1,7 @@
 import os
 import re
 import subprocess
+import sys
 import tempfile
 from pathlib import Path
 from typing import Any, Callable, Mapping, Optional
@@ -8,10 +9,17 @@ from typing import Any, Callable, Mapping, Optional
 import yaml
 
 import bento.constants
+from _pytest.python import Metafunc
+
+ALL_TESTS = [
+    d.name
+    for d in Path(__file__).parent.resolve().iterdir()
+    if d.is_dir() and (d / "commands.yaml").exists()
+]
 
 BENTO_REPO_ROOT = str(Path(__file__).parent.parent.parent.resolve())
 DYNAMIC_SECONDS = re.compile(r"([\s\S]* \d+ finding.?s.? in )\d+\.\d+( s[\s\S]*)")
-DETACHED_HEAD = re.compile(r"\[detached HEAD ([0-9a-f]+)\]")
+DETACHED_HEAD = re.compile(r"\[(detached HEAD|master) ([0-9a-f]+)\]")
 
 PIPE_OUTPUT: Mapping[str, Callable[[subprocess.CompletedProcess], str]] = {
     "expected_out": lambda r: r.stdout,
@@ -37,7 +45,7 @@ def remove_timing_seconds(string: str) -> str:
 
 
 def remove_commit_hash(string: str) -> str:
-    return re.sub(DETACHED_HEAD, r"[detached HEAD ]", string)
+    return re.sub(DETACHED_HEAD, r"[\1 ]", string)
 
 
 def match_expected(output: str, expected: str) -> bool:
@@ -53,7 +61,7 @@ def match_expected(output: str, expected: str) -> bool:
     if "finding(s) in" in expected or "findings in" in expected:
         output = remove_timing_seconds(output)
 
-    if "detached HEAD" in expected:
+    if "detached HEAD" or "master" in expected:
         output = remove_commit_hash(output)
 
     if output.strip() != expected.strip():
@@ -72,6 +80,8 @@ def check_command(step: Any, pwd: str, target: str, rewrite: bool) -> None:
     output match verification
     """
     command = step["command"]
+    if isinstance(command, str):
+        command = command.split(" ")
 
     test_identifier = f"Target:{target} Step:{step['name']}"
     env = os.environ.copy()
@@ -135,32 +145,33 @@ def run_repo(
     with open(f"tests/acceptance/{target}/commands.yaml") as file:
         info = yaml.safe_load(file)
 
-    target_repo = info["target_repo"]
-    target_hash = info["target_hash"]
+    target_repo = info.get("target_repo")
+    target_hash = info.get("target_hash")
     steps = info["steps"]
 
     with tempfile.TemporaryDirectory() as target_dir:
 
-        subprocess.run(
-            ["git", "clone", target_repo, target_dir],
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-        )
-        subprocess.run(
-            ["git", "checkout", target_hash],
-            cwd=target_dir,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-        )
-        subprocess.run(
-            ["git", "clean", "-xdf"],
-            cwd=target_dir,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            check=True,
-        )
+        if target_repo:
+            subprocess.run(
+                ["git", "clone", target_repo, target_dir],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                check=True,
+            )
+            subprocess.run(
+                ["git", "checkout", target_hash],
+                cwd=target_dir,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                check=True,
+            )
+            subprocess.run(
+                ["git", "clean", "-xdf"],
+                cwd=target_dir,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                check=True,
+            )
 
         if pre:
             pre(Path(target_dir))
@@ -169,25 +180,19 @@ def run_repo(
             check_command(step, target_dir, target, rewrite)
 
 
-# Actual Tests broken up into separate functions so progress is visible
-def test_flask() -> None:
-    run_repo("flask")
+def pytest_generate_tests(metafunc: Metafunc) -> None:
+    metafunc.parametrize("repo", ALL_TESTS)
 
 
-def test_create_react_app() -> None:
-    run_repo("create-react-app")
-
-
-def test_django_example() -> None:
-    run_repo("django-example")
-
-
-def test_instabot() -> None:
-    run_repo("instabot")
+def test_repo(repo: str) -> None:
+    run_repo(repo)
 
 
 if __name__ == "__main__":
-    run_repo("flask", rewrite=True)
-    run_repo("create-react-app", rewrite=True)
-    run_repo("django-example", rewrite=True)
-    run_repo("instabot", rewrite=True)
+    if len(sys.argv) > 1:
+        to_run = sys.argv[1:]
+    else:
+        to_run = ALL_TESTS
+
+    for t in to_run:
+        run_repo(t, rewrite=True)
